Add a module to load components from the db
This module is meant to load components from the database after we have done the
intersection routine using a list of numbers. This has two steps, first we need
to lookup all components with the given indexes, then secondly we create lists
of lists to represent each motif. We may want to use an actual data structure
for each motif in the future. The initial lookup step can be replaced in the top
level function for ease of use.

To do this I add some definitions for the db in files/schema as well as data in
files/data. We also add some simple tests for this. Only one is implemented as
a smoke test, while the rest are left failing for now.